chore: move error types and utils to api/errors.ts
