+++ site/src/api/index.ts
@@ -2,13 +2,6 @@ import axios, { AxiosRequestHeaders } from "axios"
 import { mutate } from "swr"
 import * as Types from "./types"
 
-export const Language = {
-  errorsByCode: {
-    default: "Invalid value",
-    exists: "This value is already in use",
-  },
-}
-
 const CONTENT_TYPE_JSON: AxiosRequestHeaders = {
   "Content-Type": "application/json",
 }
@@ -84,26 +77,4 @@ export const getBuildInfo = async (): Promise<Types.BuildInfoResponse> => {
 export const updateProfile = async (userId: string, data: Types.UpdateProfileRequest): Promise<Types.UserResponse> => {
   const response = await axios.put(`/api/v2/users/${userId}/profile`, data)
   return response.data
-}
-
-const getApiError = (error: unknown): Types.ApiError | undefined => {
-  if (axios.isAxiosError(error)) {
-    return error.response?.data
-  }
-}
-
-export const getFormErrorsFromApiError = (error: unknown): Record<string, string> | undefined => {
-  const apiError = getApiError(error)
-
-  if (apiError && apiError.errors) {
-    return apiError.errors.reduce((errors, error) => {
-      return {
-        ...errors,
-        [error.field]:
-          error.code in Language.errorsByCode
-            ? Language.errorsByCode[error.code as keyof typeof Language.errorsByCode]
-            : Language.errorsByCode.default,
-      }
-    }, {})
-  }
 }
